Move worker common support test func to helper_test.go file
