+++ worker/helper_test.go
@@ -17,12 +17,29 @@
 package worker
 
 import (
+	"bytes"
+	"crypto/rand"
+	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"runtime"
+	"sync"
 	"sync/atomic"
+	"time"
 
 	"github.com/CovenantSQL/CovenantSQL/blockproducer/interfaces"
+	"github.com/CovenantSQL/CovenantSQL/conf"
+	"github.com/CovenantSQL/CovenantSQL/consistent"
+	"github.com/CovenantSQL/CovenantSQL/crypto/asymmetric"
 	"github.com/CovenantSQL/CovenantSQL/crypto/hash"
+	"github.com/CovenantSQL/CovenantSQL/crypto/kms"
+	"github.com/CovenantSQL/CovenantSQL/pow/cpuminer"
 	"github.com/CovenantSQL/CovenantSQL/proto"
+	"github.com/CovenantSQL/CovenantSQL/route"
+	"github.com/CovenantSQL/CovenantSQL/rpc"
 	"github.com/CovenantSQL/CovenantSQL/types"
+	"github.com/CovenantSQL/CovenantSQL/utils"
 )
 
 var (
@@ -172,4 +189,130 @@ func (s *stubBPService) Init() {
 		profile: testOddProfiles,
 	}
 	atomic.StoreUint32(&s.count, 0)
+}
+
+func initNode() (cleanupFunc func(), server *rpc.Server, err error) {
+	var d string
+	if d, err = ioutil.TempDir("", "db_test_"); err != nil {
+		return
+	}
+
+	// init conf
+	_, testFile, _, _ := runtime.Caller(0)
+	pubKeyStoreFile := filepath.Join(d, PubKeyStorePath)
+	os.Remove(pubKeyStoreFile)
+	clientPubKeyStoreFile := filepath.Join(d, PubKeyStorePath+"_c")
+	os.Remove(clientPubKeyStoreFile)
+	dupConfFile := filepath.Join(d, "config.yaml")
+	confFile := filepath.Join(filepath.Dir(testFile), "../test/node_standalone/config.yaml")
+	if err = dupConf(confFile, dupConfFile); err != nil {
+		return
+	}
+	privateKeyPath := filepath.Join(filepath.Dir(testFile), "../test/node_standalone/private.key")
+
+	conf.GConf, _ = conf.LoadConfig(dupConfFile)
+	// reset the once
+	route.Once = sync.Once{}
+	route.InitKMS(clientPubKeyStoreFile)
+
+	var dht *route.DHTService
+
+	// init dht
+	dht, err = route.NewDHTService(pubKeyStoreFile, new(consistent.KMSStorage), true)
+	if err != nil {
+		return
+	}
+
+	// init rpc
+	if server, err = rpc.NewServerWithService(rpc.ServiceMap{route.DHTRPCName: dht}); err != nil {
+		return
+	}
+
+	// init private key
+	masterKey := []byte("")
+	if err = server.InitRPCServer(conf.GConf.ListenAddr, privateKeyPath, masterKey); err != nil {
+		return
+	}
+
+	// start server
+	go server.Serve()
+
+	cleanupFunc = func() {
+		os.RemoveAll(d)
+		server.Listener.Close()
+		server.Stop()
+		// clear the connection pool
+		rpc.GetSessionPoolInstance().Close()
+	}
+
+	return
+}
+
+// copied from sqlchain.xxx_test.
+func createRandomBlock(parent hash.Hash, isGenesis bool) (b *types.Block, err error) {
+	// Generate key pair
+	priv, pub, err := asymmetric.GenSecp256k1KeyPair()
+
+	if err != nil {
+		return
+	}
+
+	h := hash.Hash{}
+	rand.Read(h[:])
+
+	b = &types.Block{
+		SignedHeader: types.SignedHeader{
+			Header: types.Header{
+				Version:     0x01000000,
+				Producer:    proto.NodeID(h.String()),
+				GenesisHash: rootHash,
+				ParentHash:  parent,
+				Timestamp:   time.Now().UTC(),
+			},
+		},
+	}
+
+	if isGenesis {
+		// Compute nonce with public key
+		nonceCh := make(chan cpuminer.NonceInfo)
+		quitCh := make(chan struct{})
+		miner := cpuminer.NewCPUMiner(quitCh)
+		go miner.ComputeBlockNonce(cpuminer.MiningBlock{
+			Data:      pub.Serialize(),
+			NonceChan: nonceCh,
+			Stop:      nil,
+		}, cpuminer.Uint256{A: 0, B: 0, C: 0, D: 0}, 4)
+		nonce := <-nonceCh
+		close(quitCh)
+		close(nonceCh)
+		// Add public key to KMS
+		id := cpuminer.HashBlock(pub.Serialize(), nonce.Nonce)
+		b.SignedHeader.Header.Producer = proto.NodeID(id.String())
+		err = kms.SetPublicKey(proto.NodeID(id.String()), nonce.Nonce, pub)
+
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	err = b.PackAndSignBlock(priv)
+	return
+}
+
+// duplicate conf file using random new listen addr to avoid failure on concurrent test cases
+func dupConf(confFile string, newConfFile string) (err error) {
+	// replace port in confFile
+	var fileBytes []byte
+	if fileBytes, err = ioutil.ReadFile(confFile); err != nil {
+		return
+	}
+
+	var ports []int
+	if ports, err = utils.GetRandomPorts("127.0.0.1", 5000, 6000, 1); err != nil {
+		return
+	}
+
+	newConfBytes := bytes.Replace(fileBytes, []byte(":2230"), []byte(fmt.Sprintf(":%v", ports[0])), -1)
+
+	return ioutil.WriteFile(newConfFile, newConfBytes, 0644)
 }
